# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dstdelivery",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/create/category": {
            "post": {
                "tags": [
                    "Category Section"
                ],
                "summary": "Request that added a new category",
                "description": "",
                "operationId": "c93ace596ff38f36f0b8a2f731f7495e",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Please write a category title",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "photoFileName",
                        "in": "query",
                        "description": "Please write a category Photo File Name",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "big_store_id",
                        "in": "query",
                        "description": "Please write a big store id",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "tags": [
                    "User inforamtion Section"
                ],
                "summary": "Registeration",
                "description": "",
                "operationId": "30644ca260642350add4b92430618eb0",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Provide youre phone",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Provide youre name",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Provide youre email",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Provide youre password",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User inforamtion Section"
                ],
                "summary": "Log in",
                "description": "",
                "operationId": "eb0adb5292168e0c3dbee8452c98fdca",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Please write your e-mail here",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Please write your password here",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "User inforamtion Section"
                ],
                "summary": "Log out",
                "description": "",
                "operationId": "691f2638dccb469d22c8443a8120414c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/create/sub/category": {
            "post": {
                "tags": [
                    "Sub Category Section"
                ],
                "summary": "Request that added a new sub category",
                "description": "",
                "operationId": "1d9b90c17470af22232a478403c3e6ca",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Please write a category ID",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Please write a category title",
                        "required": true,
                        "allowEmptyValue": true
                    },
                    {
                        "name": "photoFileName",
                        "in": "query",
                        "description": "Please write a category Photo File Name",
                        "required": true,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/api/get/sub/catagories": {
            "get": {
                "tags": [
                    "Sub Category Section"
                ],
                "summary": "Request which returns sub categories",
                "description": "",
                "operationId": "4c1f69e229ee9f0abbe9e88501a1ae9c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "429": {
                        "description": "validation error"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}